# backend/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Weather App API
  description: |
    Complete API for Weather Application.
    
    ### Milestone 3: Full CRUD Implementation for All Entities & Open API
    
    This API provides full CRUD (Create, Read, Update, Delete) operations for:
    - Users (with authentication)
    - Cities
    - Subscriptions
    - Payments
    - Saved Filters
    
    All responses are wrapped in a consistent JSON format with error handling.
  version: 1.0.0
  contact:
    name: Weather App Support
    email: support@weatherapp.local
  license:
    name: MIT
servers:
  - url: http://localhost/weather/backend
    description: Local development server
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Get all users
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cities:
    get:
      tags: [Cities]
      summary: Get all cities
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/City'
    post:
      tags: [Cities]
      summary: Create new city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityInput'
      responses:
        '201':
          description: City created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/City'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/subscriptions:
    get:
      tags: [Subscriptions]
      summary: Get all subscriptions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
    post:
      tags: [Subscriptions]
      summary: Create new subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionInput'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/payments:
    get:
      tags: [Payments]
      summary: Get all payments
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
    post:
      tags: [Payments]
      summary: Create new payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Payment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/saved-filters:
    get:
      tags: [Saved Filters]
      summary: Get all saved filters
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavedFilter'
    post:
      tags: [Saved Filters]
      summary: Create new saved filter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedFilterInput'
      responses:
        '201':
          description: Filter created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/SavedFilter'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        fname:
          type: string
        lname:
          type: string
        email:
          type: string
        is_admin:
          type: boolean
    UserInput:
      type: object
      required: [fname, lname, email, password]
      properties:
        fname:
          type: string
          maxLength: 50
          example: John
        lname:
          type: string
          maxLength: 50
          example: Doe
        email:
          type: string
          format: email
          example: nedim@example.com
        password:
          type: string
          minLength: 6
          example: password123
        is_admin:
          type: boolean
          default: false
          example: false
    City:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        zip_code:
          type: string
        created_at:
          type: string
    CityInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        zip_code:
          type: string
    Subscription:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        created_at:
          type: string
        lasts_until:
          type: string
    SubscriptionInput:
      type: object
      required: [user_id, lasts_until]
      properties:
        user_id:
          type: integer
        lasts_until:
          type: string
    Payment:
      type: object
      properties:
        id:
          type: integer
        subscription_id:
          type: integer
        payment_method:
          type: string
        amount:
          type: number
        card_number:
          type: string
        bank_transaction_id:
          type: string
        created_at:
          type: string
    PaymentInput:
      type: object
      required: [subscription_id, payment_method, amount]
      properties:
        subscription_id:
          type: integer
        payment_method:
          type: string
        amount:
          type: number
        card_number:
          type: string
        bank_transaction_id:
          type: string
    SavedFilter:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        city_id:
          type: integer
        forecast_days:
          type: integer
        min_temp_selected:
          type: number
        max_temp_selected:
          type: number
        avg_temp_selected:
          type: number
        weather_cond_selected:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    SavedFilterInput:
      type: object
      required: [user_id, city_id]
      properties:
        user_id:
          type: integer
        city_id:
          type: integer
        forecast_days:
          type: integer
        min_temp_selected:
          type: number
        max_temp_selected:
          type: number
        avg_temp_selected:
          type: number
        weather_cond_selected:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "User not found"